+++ pkg/cps/changes.go
@@ -24,6 +24,10 @@ type (
 		CancelChange(context.Context, CancelChangeRequest) (*CancelChangeResponse, error)
 
 		// UpdateChange updates a pending change
+		// Deprecated: this function will be removed in a future release. Use one of:
+		// AcknowledgeChangeManagement(), AcknowledgePostVerificationWarnings(),
+		// AcknowledgePreVerificationWarnings(), UploadThirdPartyCertAndTrustChain()
+		// or AcknowledgeDVChallenges()
 		//
 		// See: https://techdocs.akamai.com/cps/reference/post-change-allowed-input-param
 		UpdateChange(context.Context, UpdateChangeRequest) (*UpdateChangeResponse, error)
@@ -62,7 +66,7 @@ type (
 	// Certificate is a digital certificate object
 	Certificate struct {
 		Certificate string `json:"certificate"`
-		TrustChain  string `json:"trustChain"`
+		TrustChain  string `json:"trustChain,omitempty"`
 	}
 
 	// GetChangeStatusRequest contains params required to perform GetChangeStatus
@@ -150,61 +154,63 @@ var AllowedInputContentTypeHeader = map[AllowedInputType]string{
 // Validate validates GetChangeRequest
 func (c GetChangeRequest) Validate() error {
 	return validation.Errors{
-		"enrollmentId": validation.Validate(c.EnrollmentID, validation.Required),
-		"changeId":     validation.Validate(c.ChangeID, validation.Required),
+		"EnrollmentID": validation.Validate(c.EnrollmentID, validation.Required),
+		"ChangeID":     validation.Validate(c.ChangeID, validation.Required),
 	}.Filter()
 }
 
 // Validate validates GetChangeStatusRequest
 func (c GetChangeStatusRequest) Validate() error {
 	return validation.Errors{
-		"enrollmentId": validation.Validate(c.EnrollmentID, validation.Required),
-		"changeId":     validation.Validate(c.ChangeID, validation.Required),
+		"EnrollmentID": validation.Validate(c.EnrollmentID, validation.Required),
+		"ChangeID":     validation.Validate(c.ChangeID, validation.Required),
 	}.Filter()
 }
 
 // Validate validates CancelChangeRequest
 func (c CancelChangeRequest) Validate() error {
 	return validation.Errors{
-		"enrollmentId": validation.Validate(c.EnrollmentID, validation.Required),
-		"changeId":     validation.Validate(c.ChangeID, validation.Required),
+		"EnrollmentID": validation.Validate(c.EnrollmentID, validation.Required),
+		"ChangeID":     validation.Validate(c.ChangeID, validation.Required),
 	}.Filter()
 }
 
 // Validate validates UpdateChangeRequest
 func (c UpdateChangeRequest) Validate() error {
 	return validation.Errors{
-		"enrollmentId": validation.Validate(c.EnrollmentID, validation.Required),
-		"changeId":     validation.Validate(c.ChangeID, validation.Required),
-		"allowedInputTypeParam": validation.Validate(c.AllowedInputTypeParam, validation.In(
+		"EnrollmentID": validation.Validate(c.EnrollmentID, validation.Required),
+		"ChangeID":     validation.Validate(c.ChangeID, validation.Required),
+		"AllowedInputTypeParam": validation.Validate(c.AllowedInputTypeParam, validation.In(
 			AllowedInputTypeChangeManagementACK,
 			AllowedInputTypeLetsEncryptChallengesCompleted,
 			AllowedInputTypePostVerificationWarningsACK,
 			AllowedInputTypePreVerificationWarningsACK,
 			AllowedInputTypeThirdPartyCertAndTrustChain,
 		)),
-		"certificate": validation.Validate(c.Certificate, validation.Required),
+		"Certificate": validation.Validate(c.Certificate, validation.Required),
 	}.Filter()
 }
 
 // Validate validates AcknowledgementRequest
 func (a AcknowledgementRequest) Validate() error {
 	return validation.Errors{
-		"acknowledgement": validation.Validate(a.Acknowledgement),
+		"EnrollmentID":    validation.Validate(a.EnrollmentID, validation.Required),
+		"ChangeID":        validation.Validate(a.ChangeID, validation.Required),
+		"Acknowledgement": validation.Validate(a.Acknowledgement),
 	}.Filter()
 }
 
 // Validate validates Acknowledgement
 func (a Acknowledgement) Validate() error {
 	return validation.Errors{
-		"acknowledgement": validation.Validate(a.Acknowledgement, validation.Required, validation.In(AcknowledgementAcknowledge, AcknowledgementDeny)),
+		"Acknowledgement": validation.Validate(a.Acknowledgement, validation.Required, validation.In(AcknowledgementAcknowledge, AcknowledgementDeny)),
 	}.Filter()
 }
 
 // Validate validates Certificate
 func (c Certificate) Validate() error {
 	return validation.Errors{
-		"certificate": validation.Validate(c.Certificate, validation.Required),
+		"Certificate": validation.Validate(c.Certificate, validation.Required),
 	}.Filter()
 }