+++ pkg/cps/post_verification_warnings_test.go
@@ -85,3 +85,78 @@ func TestGetChangePostVerificationWarnings(t *testing.T) {
 		})
 	}
 }
+func TestAcknowledgePostVerificationWarnings(t *testing.T) {
+	tests := map[string]struct {
+		params         AcknowledgementRequest
+		responseStatus int
+		responseBody   string
+		expectedPath   string
+		withError      func(*testing.T, error)
+	}{
+		"200 OK": {
+			params: AcknowledgementRequest{
+				EnrollmentID: 1,
+				ChangeID:     2,
+				Acknowledgement: Acknowledgement{
+					Acknowledgement: AcknowledgementAcknowledge,
+				},
+			},
+			responseStatus: http.StatusOK,
+			responseBody:   "",
+			expectedPath:   "/cps/v2/enrollments/1/changes/2/input/update/post-verification-warnings-ack",
+		},
+		"500 internal server error": {
+			params: AcknowledgementRequest{
+				EnrollmentID: 1,
+				ChangeID:     2,
+				Acknowledgement: Acknowledgement{
+					Acknowledgement: AcknowledgementAcknowledge,
+				},
+			},
+			responseStatus: http.StatusInternalServerError,
+			responseBody: `
+{
+  "type": "internal_error",
+  "title": "Internal Server Error",
+  "detail": "Error making request",
+  "status": 500
+}`,
+			expectedPath: "/cps/v2/enrollments/1/changes/2/input/update/post-verification-warnings-ack",
+			withError: func(t *testing.T, err error) {
+				want := &Error{
+					Type:       "internal_error",
+					Title:      "Internal Server Error",
+					Detail:     "Error making request",
+					StatusCode: http.StatusInternalServerError,
+				}
+				assert.True(t, errors.Is(err, want), "want: %s; got: %s", want, err)
+			},
+		},
+		"validation error": {
+			params: AcknowledgementRequest{},
+			withError: func(t *testing.T, err error) {
+				assert.True(t, errors.Is(err, ErrStructValidation), "want: %s; got: %s", ErrStructValidation, err)
+			},
+		},
+	}
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			mockServer := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+				assert.Equal(t, test.expectedPath, r.URL.String())
+				assert.Equal(t, http.MethodPost, r.Method)
+				assert.Equal(t, "application/vnd.akamai.cps.change-id.v1+json", r.Header.Get("Accept"))
+				assert.Equal(t, "application/vnd.akamai.cps.acknowledgement.v1+json", r.Header.Get("Content-Type"))
+				w.WriteHeader(test.responseStatus)
+				_, err := w.Write([]byte(test.responseBody))
+				assert.NoError(t, err)
+			}))
+			client := mockAPIClient(t, mockServer)
+			err := client.AcknowledgePostVerificationWarnings(context.Background(), test.params)
+			if test.withError != nil {
+				test.withError(t, err)
+				return
+			}
+			require.NoError(t, err)
+		})
+	}
+}